# Update perf cost expectation for test_Search_refresh