@finestructure 2022-07-14
Current (v7) + `WHERE v.latest = 'default_branch' plan stats:

<img width="332" alt="CleanShot 2022-07-14 at 11 04 35@2x" src="https://user-images.githubusercontent.com/65520/178946872-c276e165-d2e4-4e6b-9b9d-cacd189667b8.png">

<img width="321" alt="CleanShot 2022-07-14 at 11 04 24@2x" src="https://user-images.githubusercontent.com/65520/178946887-1eb07400-a80d-4bf7-9540-a43a3be23b6c.png">